# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  docs.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
      success:
        type: boolean
    type: object
...
  docs.ProfileRequest:
    properties:
      age:
        example: 30
        type: integer
      allergies:
        example:
        - '["peanuts"'
        - ' "shellfish"]'
        items:
          type: string
        type: array
      gender:
        example: male
        type: string
      goals:
        example:
        - '["weight_loss"'
        - ' "muscle_gain"]'
        items:
          type: string
        type: array
      height:
        example: 180
        type: number
      is_default:
        example: true
        type: boolean
      preferences:
        example:
        - '["vegetarian"'
        - ' "low_carb"]'
        items:
          type: string
        type: array
      weight:
        example: 75.5
        type: number
    type: object
  docs.ProfileResponse:
    properties:
      age:
        type: integer
      allergies:
        items:
          type: string
        type: array
      created_at:
        type: string
      gender:
        type: string
      goals:
        items:
          type: string
        type: array
      height:
        type: number
      id:
        type: string
      is_default:
        type: boolean
      preferences:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
      weight:
        type: number
    type: object
...
  docs.Response:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nutrimatch.com
    name: API Support
    url: http://www.nutrimatch.com/support
  description: A personalized nutrition recommendation system built on the OpenNutrition
    dataset
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: NutriMatch API
  version: "1.0"
paths:
...
  /api/v1/profiles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a profile by its ID
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete profile
      tags:
      - profiles
    get:
      consumes:
      - application/json
      description: Get a profile by its ID
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.ProfileResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get profile by ID
      tags:
      - profiles
    put:
      consumes:
      - application/json
      description: Update an existing profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated profile information
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/docs.ProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - profiles
  /api/v1/profiles/{userId}:
    post:
      consumes:
      - application/json
      description: Create a new profile for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Profile information
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/docs.ProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/docs.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      summary: Create user profile
      tags:
      - profiles
  /api/v1/profiles/user/me:
    get:
      consumes:
      - application/json
      description: Get all profiles for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/docs.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/docs.ProfileResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user profiles
      tags:
      - profiles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"